+++ src/client/cypress/e2e/helpers/testHelpers.js
@@ -1,3 +1,4 @@
+import { ObservationType } from "../../../src/pages/detail/form/hydrogeology/Observation.ts";
 import adminUser from "../../fixtures/adminUser.json";
 import editorUser from "../../fixtures/editorUser.json";
 import viewerUser from "../../fixtures/viewerUser.json";
@@ -294,6 +295,10 @@ export const getElementByDataCy = attribute => {
   return cy.get(`[data-cy=${attribute}]`);
 };
 
+export const checkElementColorByDataCy = (attribute, expectedColor) => {
+  getElementByDataCy(attribute).should("have.css", "color", expectedColor);
+};
+
 export const deleteBorehole = id => {
   cy.get("@id_token").then(token => {
     cy.request({
@@ -532,6 +537,7 @@ export const createFieldMeasurement = (
         casingId: casingId,
         fromDepthM: fromDepthM,
         toDepthM: toDepthM,
+        type: ObservationType.fieldMeasurement,
       },
       cache: "no-cache",
       credentials: "same-origin",
@@ -561,6 +567,7 @@ export const createWateringress = (
         casingId: casingId,
         fromDepthM: fromDepthM,
         toDepthM: toDepthM,
+        type: ObservationType.waterIngress,
       },
       cache: "no-cache",
       credentials: "same-origin",
@@ -590,6 +597,7 @@ export const createGroundwaterLevelMeasurement = (
         casingId: casingId,
         fromDepthM: fromDepthM,
         toDepthM: toDepthM,
+        type: ObservationType.groundwaterLevelMeasurement,
       },
       cache: "no-cache",
       credentials: "same-origin",
@@ -619,6 +627,7 @@ export const createHydrotest = (
         casingId: casingId,
         fromDepthM: fromDepthM,
         toDepthM: toDepthM,
+        type: ObservationType.hydrotest,
       },
       cache: "no-cache",
       credentials: "same-origin",